Move validate_error_trace_attributes into validators directory
