+++ tests/testing_support/fixtures.py
@@ -735,27 +735,6 @@ def validate_error_event_attributes(required_params=None, forgone_params=None, e
     return _validate_wrapper
 
 
-def validate_error_trace_attributes(err_name, required_params=None, forgone_params=None, exact_attrs=None):
-    required_params = required_params or {}
-    forgone_params = forgone_params or {}
-    exact_attrs = exact_attrs or {}
-
-    @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-    def _validate_error_trace_attributes(wrapped, instance, args, kwargs):
-        try:
-            result = wrapped(*args, **kwargs)
-        except:
-            raise
-        else:
-            target_error = core_application_stats_engine_error(err_name)
-
-            check_error_attributes(target_error.parameters, required_params, forgone_params, exact_attrs)
-
-        return result
-
-    return _validate_error_trace_attributes
-
-
 def validate_error_trace_attributes_outside_transaction(
     err_name, required_params=None, forgone_params=None, exact_attrs=None
 ):
